requirements.txt
matplotlib==3.8.4
pathlib==1.0.1
numpy==1.26.4
pandas==2.1.1
ipython==8.12.3
pillow==10.4.0